Address review comment: Functional tests for new and renamed script.
