+++ flocker/control/functional/__init__.py
@@ -1,0 +1,3 @@
+"""
+Functional tests for ``flocker.control``.
+"""
